### 2024-05-26 #6789 @editorialbot
[PRE REVIEW]: DendroPy 5: a mature Python library for phylogenetic computing 
**Submitting author:** @mmore500 (<a href="http://orcid.org/0000-0003-4726-4479">Matthew Andres Moreno</a>)
**Repository:** https://github.com/jeetsukumaran/DendroPy
**Branch with paper.md** (empty if default branch): joss
**Version:** v5.0.0
