+++ .github/workflows/ci.yaml
@@ -28,10 +28,10 @@ jobs:
       - name: Install paup
         run: |
           temp_dir=$(mktemp -d)
-          wget -P ${temp_dir} http://phylosolutions.com/paup-test/paup4a168_ubuntu64.gz
-          gunzip ${temp_dir}/paup4a168_ubuntu64.gz
-          chmod +x ${temp_dir}/paup4a168_ubuntu64
-          sudo mv ${temp_dir}/paup4a168_ubuntu64 /usr/local/bin/paup
+          wget -P ${temp_dir} http://phylosolutions.com/paup-test/paup4a169_ubuntu64.gz
+          gunzip ${temp_dir}/paup4a169_ubuntu64.gz
+          chmod +x ${temp_dir}/paup4a169_ubuntu64
+          sudo mv ${temp_dir}/paup4a169_ubuntu64 /usr/local/bin/paup
           echo "export DENDROPY_PAUP_EXECUTABLE_PATH=/usr/local/bin/paup" >> "${GITHUB_ENV}"
       - name: Install dendropy
         run: |
@@ -93,10 +93,10 @@ jobs:
       - name: Install paup
         run: |
           temp_dir=$(mktemp -d)
-          wget -P ${temp_dir} http://phylosolutions.com/paup-test/paup4a168_ubuntu64.gz
-          gunzip ${temp_dir}/paup4a168_ubuntu64.gz
-          chmod +x ${temp_dir}/paup4a168_ubuntu64
-          sudo mv ${temp_dir}/paup4a168_ubuntu64 /usr/local/bin/paup
+          wget -P ${temp_dir} http://phylosolutions.com/paup-test/paup4a169_ubuntu64.gz
+          gunzip ${temp_dir}/paup4a169_ubuntu64.gz
+          chmod +x ${temp_dir}/paup4a169_ubuntu64
+          sudo mv ${temp_dir}/paup4a169_ubuntu64 /usr/local/bin/paup
           echo "export DENDROPY_PAUP_EXECUTABLE_PATH=/usr/local/bin/paup" >> "${GITHUB_ENV}"
       - name: Unit tests with setup.py
         run: |
@@ -167,10 +167,10 @@ jobs:
       run: |
         sudo apt-get install libpython2.7  # required by paup
         temp_dir=$(mktemp -d)
-        wget -P ${temp_dir} http://phylosolutions.com/paup-test/paup4a168_ubuntu64.gz
-        gunzip ${temp_dir}/paup4a168_ubuntu64.gz
-        chmod +x ${temp_dir}/paup4a168_ubuntu64
-        sudo mv ${temp_dir}/paup4a168_ubuntu64 /usr/local/bin/paup
+        wget -P ${temp_dir} http://phylosolutions.com/paup-test/paup4a169_ubuntu64.gz
+        gunzip ${temp_dir}/paup4a169_ubuntu64.gz
+        chmod +x ${temp_dir}/paup4a169_ubuntu64
+        sudo mv ${temp_dir}/paup4a169_ubuntu64 /usr/local/bin/paup
         echo "export DENDROPY_PAUP_EXECUTABLE_PATH=/usr/local/bin/paup" >> "${GITHUB_ENV}"
     - name: Install coverage dependencies
       run: |
