### 2024-06-02 6e331bd
Install fortran for paup in CI
