+++ .github/workflows/ci.yaml
@@ -25,6 +25,8 @@ jobs:
           python-version: ${{ matrix.python-version }}
       - name: Install dependencies
         run: |
+          sudo apt-get update
+          sudo apt-get install libssl-dev
           sudo apt-get install libpython2.7 libgfortran4  # required by paup
           python -m pip install --upgrade pip
       - name: Install paup
