### 2024-06-02 1361b80
Remove sudo calls
