### 2024-06-02 7c124c3
Pass -y flag to apt commands
