+++ .github/workflows/ci.yaml
@@ -25,9 +25,9 @@ jobs:
           python-version: ${{ matrix.python-version }}
       - name: Install dependencies
         run: |
-          apt-get update
-          apt-get install libssl-dev
-          apt-get install libpython2.7 libgfortran4  # required by paup
+          apt-get update -y
+          apt-get install -y libssl-dev
+          apt-get install -y libpython2.7 libgfortran4  # required by paup
           python -m pip install --upgrade pip
       - name: Install paup
         run: |
