### 2024-06-02 c893879
Abandon container, use aptitude and ppa
