+++ .github/workflows/ci.yaml
@@ -5,8 +5,6 @@ on: [push, pull_request, workflow_dispatch]
 jobs:
   examples:
     runs-on: ubuntu-20.04
-    container:
-      image: ubuntu:18.04
     strategy:
       fail-fast: False
       matrix:
@@ -25,9 +23,9 @@ jobs:
           python-version: ${{ matrix.python-version }}
       - name: Install dependencies
         run: |
-          apt-get update -y
-          apt-get install -y libssl-dev build-essential
-          apt-get install -y libpython2.7 libgfortran4  # required by paup
+          sudo apt-get install aptitude
+          sudo aptitude install gcc-7 g++-7 gfortran-7
+          sudo apt-get install libpython2.7  # required by paup
           python -m pip install --upgrade pip
       - name: Install paup
         run: |
@@ -35,7 +33,7 @@ jobs:
           wget -P ${temp_dir} http://phylosolutions.com/paup-test/paup4a169_ubuntu64.gz
           gunzip ${temp_dir}/paup4a169_ubuntu64.gz
           chmod +x ${temp_dir}/paup4a169_ubuntu64
-          mv ${temp_dir}/paup4a169_ubuntu64 /usr/local/bin/paup
+          sudo mv ${temp_dir}/paup4a169_ubuntu64 /usr/local/bin/paup
           echo "export DENDROPY_PAUP_EXECUTABLE_PATH=/usr/local/bin/paup" >> "${GITHUB_ENV}"
       - name: Install dendropy
         run: |
