### 2024-06-02 017918a
Update rest of CI dependency installs
