+++ .github/workflows/ci.yaml
@@ -88,8 +88,9 @@ jobs:
           python-version: ${{ matrix.python-version }}
       - name: Install dependencies
         run: |
-          sudo apt-get update
-          sudo apt-get install python2 libpython2.7 libgfortran5  # required by paup
+          sudo apt-get install aptitude
+          sudo aptitude install -y gcc-7 g++-7 gfortran-7
+          sudo apt-get install libpython2.7  # required by paup
           python -m pip install --upgrade pip
           if [ -f .github/workflows/requirements.txt ]; then pip install -r .github/workflows/requirements.txt; fi
       - name: Install paup
@@ -128,8 +129,9 @@ jobs:
           python-version: ${{ matrix.python-version }}
       - name: Install dependencies
         run: |
-          sudo apt-get update
-          sudo apt-get install python2 libpython2.7 libgfortran5  # required by paup
+          sudo apt-get install aptitude
+          sudo aptitude install -y gcc-7 g++-7 gfortran-7
+          sudo apt-get install libpython2.7  # required by paup
           python -m pip install --upgrade pip
           if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
           python -m pip install .
@@ -165,9 +167,12 @@ jobs:
       uses: actions/setup-python@master
       with:
         python-version: 3.8
-    - name: Install paup
+    - name: Install dependencies
       run: |
-        sudo apt-get install libpython2.7 libgfortran5  # required by paup
+        sudo apt-get install aptitude
+        sudo aptitude install -y gcc-7 g++-7 gfortran-7
+        sudo apt-get install libpython2.7  # required by paup
+        python -m pip install --upgrade pip
         temp_dir=$(mktemp -d)
         wget -P ${temp_dir} http://phylosolutions.com/paup-test/paup4a169_ubuntu64.gz
         gunzip ${temp_dir}/paup4a169_ubuntu64.gz
