### 2024-06-02 6a7824e
Also protect curly braces for Newick
