+++ src/dendropy/dataio/newickwriter.py
@@ -379,7 +379,7 @@ class NewickWriter(ioservice.DataWriter):
             tag = nexusprocessing.escape_nexus_token(tag,
                     preserve_spaces=self.preserve_spaces,
                     quote_underscores=not self.unquoted_underscores,
-                    protect_regex=r'''[()[\],;:'"\0\t\n]''')
+                    protect_regex=r'''[()[\]{},;:'"\0\t\n]''')
             return tag
         else:
             return ""
