### 2024-06-02 b252d5b
Fix handling of space char in protect test
