+++ bumpver.toml
@@ -1,5 +1,5 @@
 [bumpver]
-current_version = "v5.0.0"
+current_version = "v5.0.1"
 version_pattern = "vMAJOR.MINOR.PATCH"
 commit_message = "bump version {old_version} -> {new_version}"
 commit = true
