### 2024-06-18 #6789 @mmore500
Yes! These potential reviewers from the joss registry could be a good fit

- Katy Brown @/katybrown
- @/tkchafin
- @/jmcbroome
- @/giraola
- @/petaripenev

In addition, we can suggest following as potential reviewers

- Michael Landis @/mlandis
- April Wright @/wrightaprilm
- Cecile Ane @/cecileane
- Claudia Solis-Lemon @/crsl4
- Tracy Heath @/trayc7
