### 2024-06-30 #6789 @tkchafin
Sure I can
-------- Original message --------From: Mahfuzur Rahman ***@***.***> Date: 30/06/2024  12:27 am  (GMT+00:00) To: openjournals/joss-reviews ***@***.***> Cc: Tyler Chafin ***@***.***>, Mention ***@***.***> Subject: Re: [openjournals/joss-reviews] [PRE REVIEW]: DendroPy 5: a mature
  Python library for phylogenetic computing  (Issue #6789)
Thanks, @petaripenev for the interest. We will start once at least two reviewers have agreed to review the article.
