### 2024-07-22 #6943 @tkchafin
@mahfuz05062 Looking through some of the initial items, I don't see the 'NonCommercial ShareAlike 4.0 International' license in the OSI list. I don't have any problems with the terms of the license in the DendroPy repo but just wanted to flag this up before checking that item off?
