### 2024-07-31 #6943 @mahfuz05062
> @mahfuz05062 I am done with my checklist. In the manuscript I didn't see an explicit section of comparing the State of the field, however there is a set of points which somewhat address this in the Statement of Need section. No direct comparisons, however I do not see the need for a direct comparison given the nature of the package. From what I saw, part of the license used for this package is BSD-3-Clause, which is an OSI approved license.

Can you point me to the place where you saw the BSD-3 clause in the license? Thanks!
