### 2024-08-01 #6943 @petaripenev
On their github in the file LICENSE.rst
https://github.com/jeetsukumaran/DendroPy/blob/main/LICENSE.rst
The first part is a one-for-one copy of the license here
https://opensource.org/license/bsd-3-clause
