### 2024-08-01 7fab22e
A couple of suggestions from Terry.
