### 2024-08-04 7d62e11
Improved docstring for src/dendropy/datamodel/treemodel/_tree.py/reroot_at_edge
