### 2024-08-05 725d4ac
Removing dorder function.
