### 2024-08-05 0cc1255
Removed construction of order dict in polytomy default order sorting.
