### 2024-08-06 d4d9200
Update src/dendropy/datamodel/treemodel/_tree.py
