+++ src/dendropy/datamodel/treemodel/_tree.py
@@ -2590,7 +2590,7 @@ class Tree(
         the tree on the new node.
         ``length1`` will be assigned to the new (sub-)edge leading
         to the original parent of the original edge. ``length2`` will be
-        assigned to new (sub-)edge leading to the original child of the original edge.
+        assigned to the new (sub-)edge leading to the original child of the original edge.
         If ``update_bipartitions`` is True, then the edges' ``bipartition`` and the tree's
         ``bipartition_encoding`` attributes will be updated.
         If the *old* root of the tree had an outdegree of 2, then after this
