### 2024-08-06 9be4185
Create LEGAL_NOTICES.rst
