+++ LEGAL_NOTICES.rst
@@ -1,0 +1,45 @@
+1.  If you use this work or any portion thereof in published work
+    please cite it as:
+
+        Moreno, M. A., Sukumaran, J., and M. T. Holder. 2024.
+        DendroPy 5: a mature Python library for phylogenetic computing.
+        arXiv preprint arXiv:2405.14120. https://doi.org/10.48550/arXiv.2405.14120
+
+2.  All DendroPy documentation, (non-logo) images and other media, as
+    provided online or otherwise distributed, can be used in accordance with the
+    `Attribution-NonCommercial-ShareAlike 4.0 International License
+    <http://creativecommons.org/licenses/by-nc-sa/4.0/>`_ , which allows anyone
+    to copy, modify and redistribute modifications of all or part of the DendroPy
+    documentation as long as:
+
+        * The license is included with all copies or redistributions.
+        * the DendroPy documentation is attributed as the originating document.
+        * You must give appropriate credit, provide a link to the license, and
+            indicate if changes were made. You may do so in any reasonable
+            manner, but not in any way that suggests the licensor endorses you or
+            your use.
+        * You may not use the material for commercial purposes.
+        * If you remix, transform, or build upon the material, you must
+            distribute your contributions under the same license as the original.
+
+3.  The DendroPy logo can be used and reproduced freely only to identify
+    advertise, or otherwise denote association or usage of the DendroPy
+    library.
+
+4.  The test data files (available at ``dendropy/test/data``) have all been (a)
+    sourced from public domain data, (b) obtained as part of example data
+    bundled with other applications, and typically having been used in
+    published analyses (c) generated using analysis programs from public domain
+    data.
+
+    In the case of (a), the data is from GenBank, and thus in public domain.
+    Where possible, we have maintained citation information for the original
+    authors/work that generated the data.
+
+    In the case of (b), the original data is also available from the public
+    domain. Where possible, we have maintained citation information for the
+    original authors/work that generated the data.
+
+    In the case of (c), no program we have used places any restriction on their
+    output. We maintain the generated text ("File generated by ... etc")
+    to track provenance.
