### 2024-08-06 #6943 @mmore500
I’ve opened a PR that sets our license to boilerplate BSD. I am still clearing this with my collaborators, but we will be sure to resolve this licensing issue soon. I will update this thread when we make a decision.

https://github.com/jeetsukumaran/DendroPy/pull/207

In the meantime, I hope that the review can proceed in other regards!
