### 2024-08-06 173a9b2
Separate reorder functionality to own method
