### 2024-08-07 #6943 @mmore500
Quick update, we completed the process of switching to a standard BSD 3 clause license :+1:
