### 2024-08-15 #6943 @tkchafin
Finished my checklist. The only suggestions I would have would be:
1. Update the installation instructions to also include the Conda option -- also maybe remove sudo from the default command
2. Add links in README to the documentation on issue reporting and contribution guidelines so these are easier to find
3. Since this is the first published update in a while, it might be good to summarise the major feature differences in the paper since the 2010 paper?
