### 2024-08-18 2f245c9
Remove sudo from pip commands
RE https://github.com/openjournals/joss-reviews/issues/6943#issuecomment-2290941927
