### 2024-08-18 8343a4e
Add README info on contributing/issue tracker
RE https://github.com/openjournals/joss-reviews/issues/6943#issuecomment-2290941927
