### 2024-08-18 #6943 @mmore500
- [ ] Update the installation instructions to also include the Conda option
- [ ] also maybe remove sudo from the default command
- [ ] Add links in README to the documentation on issue reporting and contribution guidelines so these are easier to find
Since this is the first published update in a while, it might be good to summarise the major feature differences in the paper since the 2010 paper?
