### 2024-08-18 0a188b3
Add conda install instructions
RE https://github.com/openjournals/joss-reviews/issues/6943#issuecomment-2290941927
