### 2024-08-18 327a712
Note changes since original dendropy release
RE https://github.com/openjournals/joss-reviews/issues/6943#issuecomment-2290941927
