### 2024-08-18 2884a24
Update MIDAS and Schmidt info
