+++ src/dendropy/model/birthdeath.py
@@ -194,28 +194,28 @@ def birth_death_tree(birth_rate, death_rate, birth_rate_sd=0.0, death_rate_sd=0.
             and ("num_extinct_tips" not in kwargs)
             and ("num_total_tips" not in kwargs)
             and ("max_time" not in kwargs) ):
-        if "taxon_namespace" in kwargs:
-            ### cannot support legacy approach, b/c ``taxon_namespace`` may grow during function, leading to unpredictable behavior
-            # deprecate.dendropy_deprecation_warning(
-            #         preamble="Deprecated: The 'taxon_namespace' argument can no longer be used to specify a termination condition as a side-effect. Use one or more of the following instead with the length of the taxon namespace instance as a value: 'num_extant_tips', 'num_extinct_tips', or 'num_total_tips'",
-            #         old_construct="tree = birth_death_tree(\n    ...\n    taxon_namespace=taxon_namespace,\n    ...\n)",
-            #         new_construct="tree = birth_death_tree(\n    ...\n    taxon_namespace=taxon_namespace,\n    num_extant_tips=len(taxon_namespace),\n    ...\n)")
-            # kwargs["num_extant_tips"] = len(kwargs["taxon_namespace"])
-            raise ValueError("The 'taxon_namespace' argument can no longer be used to specify a termination condition as a side-effect."
-                             "Use one or more of the following instead with the length of the taxon namespace instance as a value: "
-                             "'num_extant_tips', 'num_extinct_tips', or 'num_total_tips'.\n"
-                             "That is, instead of:\n\n"
-                             "    tree = birth_death_tree(\n        ...\n        taxon_namespace=taxon_namespace,\n        ...\n    )\n\n"
-                             "Use:\n\n"
-                             "    ntax = len(taxon_namespace)\n    tree = birth_death_tree(\n        ...\n        taxon_namespace=taxon_namespace,\n        num_extant_tips=ntax,\n        ...\n    )\n"
-                             "\nOr (recommended):\n\n"
-                             "    tree = birth_death_tree(\n        ...\n        taxon_namespace=taxon_namespace,\n        num_extant_tips=100,\n        ...\n    )\n"
-                             "\nNote that the taxon namespace instance size may grow during any particular call of the function depending on taxon assignment/creation settings, so"
-                             " for stable and predictable behavor it is important to take a snapshot of the desired taxon namespace size before any call of the function, or, better yet"
-                             " simply pass in a constant value."
-                             )
-        else:
-            raise ValueError("One or more of the following must be specified: 'num_extant_tips', 'num_extinct_tips', or 'max_time'")
+        # if "taxon_namespace" in kwargs:
+        #     ### cannot support legacy approach, b/c ``taxon_namespace`` may grow during function, leading to unpredictable behavior
+        #     # deprecate.dendropy_deprecation_warning(
+        #     #         preamble="Deprecated: The 'taxon_namespace' argument can no longer be used to specify a termination condition as a side-effect. Use one or more of the following instead with the length of the taxon namespace instance as a value: 'num_extant_tips', 'num_extinct_tips', or 'num_total_tips'",
+        #     #         old_construct="tree = birth_death_tree(\n    ...\n    taxon_namespace=taxon_namespace,\n    ...\n)",
+        #     #         new_construct="tree = birth_death_tree(\n    ...\n    taxon_namespace=taxon_namespace,\n    num_extant_tips=len(taxon_namespace),\n    ...\n)")
+        #     # kwargs["num_extant_tips"] = len(kwargs["taxon_namespace"])
+        #     raise ValueError("The 'taxon_namespace' argument can no longer be used to specify a termination condition as a side-effect."
[…]
