+++ src/dendropy/simulate/treesim.py
@@ -101,3 +101,4 @@ def mapped_birthdeath_trees(
 ):
     return [*map_birthdeath_trees(fn, rng, model_kwargs, n_replicates)]
 
+
