### 2024-08-24 #6943 @mahfuz05062
Hi @mmore500,

I have some comments on the text of the article:

- Summary: can you please clarify/rewrite the first line of summary for clarity? Also, for the third line, "Insights from ... have" would be more appropriate in my opinion.
- In Statement of need, point 4, you should also start with a capital for consistency.
- Please use ',' and ';' consistently across your bullet points and numberings. For example, you have a mixed use of ',' and ';' for bullet points in Features. Fix this wherever appropriate.
- In Features, please start with "Key features of DendroPy are:".
- Please rewrite the first line of maintenance for clarity.
- Please use "includes" instead of "included" in the first line of last paragraph of Maintenance as present tense has been used throughout. The last line ("we look forward ...") of this paragraph seems unnecessary.
