### 2024-08-24 d56b852
Revise third line and fix inconsistent list punctuation
