### 2024-08-24 d8c7ef5
Fix tenses and phrasing
