### 2024-08-24 ad536e5
Strip superfluous closing sentence
