### 2024-08-24 6db11ab
Revise verbose sentence opening Maintenance
