### 2024-08-24 18f0a49
Simplify opening sentence
