### 2024-08-27 #6943 @mahfuz05062
Hi @mmore500,

Thank you for making the changes. A couple more requests:
- Can you add your self-citation (in the first paragraph of the summary) inside parentheses and highlight "Moreno et al. 2023" (also use a comma like the others) for consistency?
- I would rewrite the first paragraph of Maintenance as: "The primary focus of DenroPy's version 5 release is to support sustainable long-term maintenance by reducing the effort needed for new releases. We hope this will result in a regular release schedule ...". Also, could you add a line about why/how this release reduces the maintenance burden to facilitate future releases?
