### 2024-08-30 86dd12a
Fix citation syntax
