+++ joss/paper.bib
@@ -406,7 +406,7 @@
   number = 5,
   pages = {1154--1170},
   doi = {10.1093/sysbio/syad040},
-  issn = {1076-836x},
+  issn = {1063-5157},
   url = {http://dx.doi.org/10.1093/sysbio/syad040},
   editor = {Gascuel,  Olivier},
 }
@@ -551,7 +551,6 @@
   volume = {00},
   number = 1,
   doi = {10.1002/0471250953.bi0604s00},
-  issn = {1934-340x},
   url = {http://dx.doi.org/10.1002/0471250953.bi0604s00},
 }
 @article{redelings2017supertree,
@@ -585,7 +584,7 @@
 @article{sukumaran2015machine,
   title = {Machine Learning Biogeographic Processes from Biotic Patterns: A New Trait-Dependent Dispersal and Diversification Model with Model Choice By Simulation-Trained Discriminant Analysis},
   volume = {65},
-  ISSN = {1076-836X},
+  ISSN = {1063-5157},
   url = {http://dx.doi.org/10.1093/sysbio/syv121},
   DOI = {10.1093/sysbio/syv121},
   number = {3},
