### 2024-09-05 #6943 @mmore500
Hi @mahfuz05062 --- I just pushed a change to the DendroPy `joss` branch that should fix this issue. Could you try compiling the manuscript again from your end and see if it does the trick?
