### 2024-09-06 ff79b68
Fix Knowles ISSN
