+++ joss/paper.bib
@@ -232,7 +232,7 @@
   number = 12,
   pages = {2623--2635},
   doi = {10.1046/j.1365-294x.2002.01410.x},
-  issn = {1365-294x},
+  issn = {1545-2069},
   url = {http://dx.doi.org/10.1046/j.1365-294X.2002.01410.x},
 }
 @inbook{lalejini2024phylogeny,
