### 2024-09-06 8bd7ae9
Strip Lewison ISSN
