+++ joss/paper.bib
@@ -272,7 +272,6 @@
   number = 4,
   pages = {581--596},
   doi = {10.1038/s41559-023-02000-4},
-  issn = {2397-334x},
   url = {http://dx.doi.org/10.1038/s41559-023-02000-4},
 }
 @article{lipman1985rapid,
