### 2024-09-06 #6943 @mmore500
It looks like I somehow missed two remaining ISSN’s ending in x. I fixed the two listed and read through the bib file manually to check for any more. Could you try again?
