### 2024-09-16 #6943 @Kevin-Mattheus-Moerman
@mmore500
as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. I will now process these final checks:

- [ ] Review completed, e.g. all review tick boxes ticked.
- [ ] Software license tag listed here matches a tagged release

- [ ] Checked paper formatting, references, and typos

- [ ] In your affiliations, please spell out USA as United States of America.
