### 2024-09-16 #6943 @mmore500
Hi @Kevin-Mattheus-Moerman — Thanks for looking over our manuscript! I have taken care of the remaining point you identified in https://github.com/jeetsukumaran/DendroPy/pull/213/commits/ff390ab94e064eacbae1655ef6b31f0f082208e1 👍
